refactor: rename lib to util (clearer, the lib is going to contain all the particle stuff)
